Using SFML 2.3 for example, added scrolling to control list and combo box popup.

+++ Source/DatGui.cpp
@@ -319,7 +319,7 @@ bool DatGui::onMouseMove(int x, int y)
 	return gui->onMouseMove(x, y);
 }
 
-bool DatGui::onMouseWheel(int x, int y, int dy)
+bool DatGui::onMouseWheel(int x, int y, float dy)
 {
 	if(gui == NULL)
 	{
